securelaunch: revert commit 0128a29c: bring back sdX support.
that has long been an issue and people should really use UUIDs but it is still
legit to specify sdX, so we shouldn't force users if they don't want to convert
The view I take is that as long as the kernel accepts it, we should we should
not limit end users over our personal views. I would accept deprecation if
u-root took that position.I would let them know I disagree but would fall into
compliance since they are upstream.

it doesn't hurt to support both but if a user should be ready for the pain.