securelaunch: revert 0f4ec61d: no need to segregate kexec load from kexec reboot
We can revisit this later since we are either way going to change the kexec
system call eventually. this will be done because the new kexec call developed
by matthew garret will have kexec doing all measurements inside the kernel as
part of the kexec system call. when that occurs, we wont need to measure
kernel, initrd inside u-root.