securelaunch: helpers, sluinit: PersistData Queue to defer all persist actions to last step of sluinit.
If a device is already mounted(say with read flag), but we now need to
mount it with write flag, a remount(mount + unmount) operation is needed.

sdX-Y means: sdX is the device, Y is the flag(0 for write, 1 for read)...
Example:
sda1-0
sda1-1 // remount operation
sda1-0 // remount operation
sda1-1 // remount operation
Total re-mount operations: 3

To reduce the number of remounts in sluinit, this commit defers all write flag mount
operations to end of sluinit just before kexec.
With this change, we have sequence

sda1-1
sda1-1
sda1-0 // remount operation
sda1-0

Total re-mount operations: 1

This change has been implemented using a simple persist data queue,
which is a slice to hold data that needs to be persisted to disk,

Signed-off-by: Simran Singh <simran.p.singh@oracle.com>